feat: show progress in operator
